+++ packages/expo-modules-core/ios/Swift/Exceptions/UnexpectedException.swift
@@ -6,8 +6,9 @@
 public class UnexpectedException: Exception {
   private let errorDescription: String
 
-  public init(_ error: Error) {
+  public init(_ error: Error, file: String = #fileID, line: UInt = #line, function: String = #function) {
     self.errorDescription = error.localizedDescription
+    super.init(file: file, line: line, function: function)
   }
 
   public override var reason: String {