+++ packages/expo-modules-core/ios/Swift/ModuleHolder.swift
@@ -63,47 +63,22 @@ public final class ModuleHolder {
 
   // MARK: Calling functions
 
-  func call(function functionName: String, args: [Any], promise: Promise) {
-    do {
-      guard let function = definition.functions[functionName] else {
-        throw FunctionNotFoundException((functionName: functionName, moduleName: self.name))
-      }
-      let queue = function.queue ?? DispatchQueue.global(qos: .default)
-
-      // Given arguments can be:
-      // - Swift primitives when invoked through the bridge and in unit tests
-      // - `JavaScriptValue`s when the function is called through the JSI
-      // The latter need to be unpacked to Swift primitives on the JS thread,
-      // so we do the casting before the function call is scheduled on the queue.
-      let arguments = try castArguments(args, toTypes: function.argumentTypes)
-
-      queue.async {
-        function.call(args: arguments, promise: promise)
-      }
-    } catch let error as CodedError {
-      promise.reject(error)
-    } catch {
-      promise.reject(UnexpectedException(error))
+  func call(function functionName: String, args: [Any], _ callback: @escaping (FunctionCallResult) -> () = { _ in }) {
+    guard let function = definition.functions[functionName] else {
+      callback(.failure(FunctionNotFoundException((functionName: functionName, moduleName: self.name))))
+      return
     }
-  }
-
-  func call(function functionName: String, args: [Any], _ callback: @escaping (Any?, CodedError?) -> Void = { _, _ in }) {
-    let promise = Promise {
-      callback($0, nil)
-    } rejecter: {
-      callback(nil, $0)
-    }
-    call(function: functionName, args: args, promise: promise)
+    function.call(args: args, callback: callback)
   }
 
   @discardableResult
   func callSync(function functionName: String, args: [Any]) -> Any? {
-    guard let function = definition.functions[functionName] else {
+    guard let function = definition.functions[functionName] as? AnySyncFunctionComponent else {
       return nil
     }
     do {
       let arguments = try castArguments(args, toTypes: function.argumentTypes)
-      return function.callSync(args: arguments)
+      return try function.call(args: arguments)
     } catch {
       return error
     }
@@ -119,9 +94,10 @@ public final class ModuleHolder {
    */
   private func createJavaScriptModuleObject() -> JavaScriptObject? {
     // It might be impossible to create any object at the moment (e.g. remote debugging, app context destroyed)
-    guard let object = appContext?.runtime?.createObject() else {
+    guard let runtime = appContext?.runtime else {
       return nil
     }
+    let object = runtime.createObject()
 
     // Fill in with constants
     for (key, value) in getConstants() {
@@ -129,11 +105,11 @@ public final class ModuleHolder {
     }
 
     // Fill in with functions
-    for (_, fn) in definition.functions {
-      if fn.isAsync {
-        object.setAsyncFunction(fn.name, argsCount: fn.argumentsCount, block: createAsyncFunctionBlock(holder: self, name: fn.name))
+    for fn in definition.functions.values {
+      if fn is AnyAsyncFunctionComponent {
+        object.setProperty(fn.name, value: runtime.createAsyncFunction(fn.name, argsCount: fn.argumentsCount, block: createAsyncFunctionBlock(holder: self, name: fn.name)))
       } else {
-        object.setSyncFunction(fn.name, argsCount: fn.argumentsCount, block: createSyncFunctionBlock(holder: self, name: fn.name))
+        object.setProperty(fn.name, value: runtime.createSyncFunction(fn.name, argsCount: fn.argumentsCount, block: createSyncFunctionBlock(holder: self, name: fn.name)))
       }
     }
     return object
@@ -166,9 +142,9 @@ public final class ModuleHolder {
    */
   func modifyListenersCount(_ count: Int) {
     if count > 0 && listenersCount == 0 {
-      _ = definition.functions["startObserving"]?.callSync(args: [])
+      definition.functions["startObserving"]?.call(args: [])
     } else if count < 0 && listenersCount + count <= 0 {
-      _ = definition.functions["stopObserving"]?.callSync(args: [])
+      definition.functions["stopObserving"]?.call(args: [])
     }
     listenersCount = max(0, listenersCount + count)
   }