+++ packages/expo-modules-core/ios/Swift/Objects/ObjectDefinitionComponents.swift
@@ -42,8 +42,8 @@ public func Constants(_ body: @autoclosure @escaping () -> [String: Any?]) -> An
 public func function<R>(
   _ name: String,
   _ closure: @escaping () throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [],
     closure
@@ -57,8 +57,8 @@ public func function<R>(
 public func function<R, A0: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [ArgumentType(A0.self)],
     closure
@@ -72,8 +72,8 @@ public func function<R, A0: AnyArgument>(
 public func function<R, A0: AnyArgument, A1: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [ArgumentType(A0.self), ArgumentType(A1.self)],
     closure
@@ -87,8 +87,8 @@ public func function<R, A0: AnyArgument, A1: AnyArgument>(
 public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1, A2) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [
       ArgumentType(A0.self),
@@ -106,8 +106,8 @@ public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument>(
 public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1, A2, A3) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [
       ArgumentType(A0.self),
@@ -126,8 +126,8 @@ public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: A
 public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: AnyArgument, A4: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1, A2, A3, A4) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [
       ArgumentType(A0.self),
@@ -147,8 +147,8 @@ public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: A
 public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: AnyArgument, A4: AnyArgument, A5: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1, A2, A3, A4, A5) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [
       ArgumentType(A0.self),
@@ -169,8 +169,8 @@ public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: A
 public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: AnyArgument, A4: AnyArgument, A5: AnyArgument, A6: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1, A2, A3, A4, A5, A6) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [
       ArgumentType(A0.self),
@@ -192,8 +192,8 @@ public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: A
 public func function<R, A0: AnyArgument, A1: AnyArgument, A2: AnyArgument, A3: AnyArgument, A4: AnyArgument, A5: AnyArgument, A6: AnyArgument, A7: AnyArgument>(
   _ name: String,
   _ closure: @escaping (A0, A1, A2, A3, A4, A5, A6, A7) throws -> R
-) -> AnyFunction {
-  return ConcreteFunction(
+) -> AnyAsyncFunctionComponent {
+  return AsyncFunctionComponent(
     name,
     argTypes: [
       ArgumentType(A0.self),
@@ -231,14 +231,14 @@ public func Events(_ names: String...) -> AnyDefinition {
  */
 @available(*, deprecated, renamed: "OnStartObserving")
 public func onStartObserving(_ body: @escaping () -> Void) -> AnyFunction {
-  return ConcreteFunction("startObserving", argTypes: [], body)
+  return AsyncFunctionComponent("startObserving", argTypes: [], body)
 }
 
 /**
  Function that is invoked when the first event listener is added.
  */
 public func OnStartObserving(_ body: @escaping () -> Void) -> AnyFunction {
-  return ConcreteFunction("startObserving", argTypes: [], body)
+  return AsyncFunctionComponent("startObserving", argTypes: [], body)
 }
 
 /**
@@ -246,12 +246,12 @@ public func OnStartObserving(_ body: @escaping () -> Void) -> AnyFunction {
  */
 @available(*, deprecated, renamed: "OnStopObserving")
 public func onStopObserving(_ body: @escaping () -> Void) -> AnyFunction {
-  return ConcreteFunction("stopObserving", argTypes: [], body)
+  return AsyncFunctionComponent("stopObserving", argTypes: [], body)
 }
 
 /**
  Function that is invoked when all event listeners are removed.
  */
 public func OnStopObserving(_ body: @escaping () -> Void) -> AnyFunction {
-  return ConcreteFunction("stopObserving", argTypes: [], body)
+  return AsyncFunctionComponent("stopObserving", argTypes: [], body)
 }